Add loading data from api

+++ app/src/main/java/my/github/dstories/model/ShortMonster.kt
@@ -7,5 +7,5 @@ data class ShortMonster(
     val portrait: ImagePath?,
     val hitPoints: Int,
     val armorClass: Int,
-    val challengeRating: Int
+    val challengeRating: Double
 )
